reorg the code and add checkstyle and findbugs
