+++ build.gradle
@@ -1,24 +1,39 @@
-plugins {
-    id 'java'
-}
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
 
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-
+allprojects {
+    apply plugin: 'idea'
+    apply plugin: 'eclipse'
+    if (file("src/main/java").isDirectory())
+    {
+        apply plugin: 'java'
+    }
+    apply from: "$rootDir/gradle/checkstyle.gradle"
+    apply from: "$rootDir/gradle/findbugs.gradle"
+    apply from: "$rootDir/gradle/java.gradle"
+    repositories {
+        jcenter()
+        mavenCentral()
+        mavenLocal()
+    }
+    version = '1.0-SNAPSHOT'
+    group = "io.pravega"
+    sourceCompatibility = 1.8
 }
 
 dependencies {
     testCompile 'junit:junit:4.12'
-
-    compile "io.pravega:pravega-client:0.4.0-1944.dc5d241-SNAPSHOT",
-            "io.pravega:pravega-common:0.4.0-1944.dc5d241-SNAPSHOT"
-
     compile 'commons-cli:commons-cli:1.4'
-
     compile 'org.slf4j:slf4j-simple:1.7.9'
+    compile 'io.pravega:pravega-client:0.3.2'
+    compile 'io.pravega:pravega-common:0.3.2'
+
 }
