+++ checkstyle/checkstyle.xml
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC
+    "-//Puppy Crawl//DTD Check Configuration 1.3//EN"
+     "http://www.puppycrawl.com/dtds/configuration_1_3.dtd">
+<module name="Checker">
+  <property name="localeLanguage" value="en"/>
+
+  <module name="FileTabCharacter"/>
+
+  <!-- header -->
+  <module name="RegexpHeader">
+    <property name="header" value="/\*\*\n *"/>
+  </module>
+  <module name="SuppressionFilter">
+        <property name="file" value="checkstyle/suppressions.xml"/>
+  </module>
+  <module name="SuppressWarningsFilter"/>
+  <module name="FileTabCharacter"/>
+  <module name="TreeWalker">
+    <module name="SuppressWarningsHolder" />
+    <module name="SuppressWarnings"/>
+    <!-- code cleanup -->
+    <module name="UnusedImports"/>
+    <module name="RedundantImport"/>
+    <module name="IllegalImport" />
+    <module name="EqualsHashCode"/>
+    <module name="NoClone"/>
+    <module name="MissingOverride"/>
+    <module name="IllegalThrows">
+        <property name="illegalClassNames" value="java.lang.Throwable, java.lang.Error, java.lang.RuntimeException, PravegaException"/>
+    </module>
+    <module name="FallThrough"/>
+    <module name="SimplifyBooleanExpression"/>
+    <module name="OneStatementPerLine"/>
+    <module name="UnnecessaryParentheses" />
+    <module name="SimplifyBooleanReturn"/>
+    <module name="DeclarationOrder"/>
+    <!--<module name="InnerAssignment"/> This doesn't work with lambdas-->
+    <module name="ModifiedControlVariable"/>
+    <module name="OverloadMethodsDeclarationOrder"/>
+    <module name="ReturnCount">
+        <property name="maxForVoid" value="4"/>
+        <property name="max" value="4"/>
+    </module>
+    <module name="StringLiteralEquality"/>
+    
+
+    <!-- style -->
+    <module name="NoLineWrap"/>
+    <module name="DefaultComesLast"/>
+    <module name="EmptyStatement"/>
+    <module name="ArrayTypeStyle"/>
+    <module name="UpperEll"/>
+    <module name="LeftCurly"/>
+    <module name="RightCurly"/>
+    <module name="ConstantName">
+      <property name="format" value="(^[A-Z][A-Z0-9]*(_[A-Z0-9]+)*$)|(^log$)"/>
+    </module>
+    <module name="LocalVariableName"/>
+    <module name="LocalFinalVariableName"/>
+    <module name="MemberName"/>
+    <module name="ClassTypeParameterName">
+        <!-- Allows type parameters that are a single capital letter (with optional number) or a name that either begins on ends with 'T'
+        IE: T, E, E1, TFoo, or FooT -->
+      <property name="format" value="(^[A-Z][0-9]?)$|([A-Z][a-zA-Z0-9]*[T]$)|(T[a-zA-Z0-9]*$)"/>
+    </module>
+    <module name="MethodTypeParameterName">
+        <!-- Allows type parameters that are a single capital letter (with optional number) or a name that either begins on ends with 'T'
+        IE: T, E, E1, TFoo, or FooT -->
+      <property name="format" value="(^[A-Z][0-9]?)$|([A-Z][a-zA-Z0-9]*[T]$)|(T[a-zA-Z0-9]*$)"/>
+    </module>
+    <module name="PackageName"/>
+    <module name="ParameterName"/>
+    <module name="StaticVariableName">
+        <property name="format" value="Static non-final variables are not allowed."/>
+    </module>
+    <module name="TypeName"/>
+    <module name="AvoidStarImport">
+       <!--    <property name="excludes" value="java.io,java.net,java.lang.Math"/> -->
+           <property name="allowClassImports" value="false"/>
+           <property name="allowStaticMemberImports" value="true"/>
+    </module>
+
+    <!-- dependencies -->
+    <module name="ImportControl">
+      <property name="file" value="${importControlFile}"/>
+    </module>
+
+    <!-- whitespace -->
+    <module name="GenericWhitespace"/>
+    <module name="NoWhitespaceBefore"/>
+    <module name="WhitespaceAfter" />
+    <module name="NoWhitespaceAfter">
+        <property name="tokens" value="INC, DEC, UNARY_MINUS, UNARY_PLUS, BNOT, LNOT, DOT, ARRAY_DECLARATOR, INDEX_OP"/>
+    </module>
+    <module name="WhitespaceAround">
+      <property name="allowEmptyConstructors" value="true"/>
+      <property name="allowEmptyMethods" value="true"/>
+      <property name="tokens" value="ASSIGN, BAND, BAND_ASSIGN, BOR, BOR_ASSIGN, BSR, BSR_ASSIGN, BXOR, BXOR_ASSIGN, COLON, DIV, DIV_ASSIGN, DO_WHILE, EQUAL, GE, GT, LAMBDA, LAND, LCURLY, LE, LITERAL_CATCH, LITERAL_DO, LITERAL_ELSE, LITERAL_FINALLY, LITERAL_FOR, LITERAL_IF, LITERAL_RETURN, LITERAL_SWITCH, LITERAL_SYNCHRONIZED, LITERAL_TRY, LITERAL_WHILE, LOR, LT, MINUS_ASSIGN, MOD, MOD_ASSIGN, NOT_EQUAL, PLUS_ASSIGN, QUESTION, RCURLY, SL, SLIST, SL_ASSIGN, SR, SR_ASSIGN, STAR, STAR_ASSIGN, LITERAL_ASSERT, TYPE_EXTENSION_AND"/>
+    </module>
+    <!--<module name="Indentation"/> TODO: this needs to be re-enabled once we figure out the lambda indentation problem. -->
+    <module name="MethodParamPad"/>
+    <module name="TypecastParenPad"/>
+    <module name="NeedBraces"/>
+    <module name="CommentsIndentation"/>
+    <module name="EmptyLineSeparator">
+      <property name="tokens" value="METHOD_DEF, CTOR_DEF, INTERFACE_DEF, CLASS_DEF, ENUM_DEF"/>
+      <property name="allowMultipleEmptyLinesInsideClassMembers" value="false"/>
+    </module>
+    <module name="AvoidNestedBlocks"/>
+    <module name="JavadocStyle">
+      <property name="scope" value="protected"/>
+      <property name="checkFirstSentence" value="true"/>
+      <property name="checkEmptyJavadoc" value="true"/>
+      <property name="checkHtml" value="true"/>
+    </module>
+    <module name="MissingOverride"/>
+    <module name="MissingDeprecated"/>
+    <module name="AnnotationLocation">
+        <property name="allowSamelineMultipleAnnotations" value="false"/>
+        <property name="allowSamelineSingleParameterlessAnnotation" value="false"/>
+        <property name="allowSamelineParameterizedAnnotation" value="false"/>
+        <property name="tokens" value="VARIABLE_DEF, PARAMETER_DEF"/>
+    </module>
+    <module name="JavadocMethod">
+      <property name="scope" value="protected"/>
+      <property name="validateThrows" value="true"/>
+      <property name="allowUndeclaredRTE" value="true"/>
+      <property name="allowMissingThrowsTags" value="false"/>
+      <property name="allowMissingParamTags" value="false"/>
+      <property name="allowMissingReturnTag" value="true"/>
+      <property name="allowMissingJavadoc" value="true"/> <!--TODO: this should be enabled at one point. -->
+      <property name="ignoreMethodNamesRegex" value="^get.*$"/> <!--It would be nice if we could allow checking that the doc exists without also verifying the @returns is there also, but checkstyle does not allow that. -->
+      <property name="suppressLoadErrors" value="true"/>
+    </module>
+  </module>
+</module>
