+++ checkstyle/example.importorder
@@ -1,0 +1,4 @@
+#Organize Import Order
+#Thu Apr 06 14:13:16 PDT 2017
+1=\#
+0=
