+++ checkstyle/import-control.xml
@@ -1,0 +1,49 @@
+<!DOCTYPE import-control PUBLIC
+        "-//Puppy Crawl//DTD Import Control 1.1//EN"
+        "http://www.puppycrawl.com/dtds/import_control_1_1.dtd">
+<import-control pkg="io.pravega">
+
+    <!-- THINK HARD ABOUT THE LAYERING OF THE PROJECT BEFORE CHANGING THIS FILE -->
+
+    <!-- common library dependencies -->
+    <allow pkg="java"/>
+    <allow pkg="javax.management"/>
+    <allow pkg="javax.annotation"/>
+    <allow pkg="org.apache.commons.lang"/>
+    <allow pkg="org.slf4j"/>
+    <allow pkg="org.junit"/>
+    <allow pkg="org.mockito"/>
+    <allow pkg="java.security"/>
+    <allow pkg="javax.net.ssl"/>
+    <allow pkg="javax.security"/>
+    <allow pkg="org.ietf.jgss"/>
+    <allow pkg="com.google"/>
+    <allow pkg="io.pravega"/>
+    <allow pkg="lombok"/>
+    <allow pkg="ch.qos.logback"/>
+    <allow pkg="io.netty"/>
+    <allow pkg="org.apache"/>
+    <allow pkg="com.typesafe.config"/>
+    <allow pkg="org.rocksdb"/>
+    <allow pkg="com.codahale"/>
+    <allow pkg="com.emc.object"/>
+    <allow pkg="com.readytalk"/>
+    <allow pkg="com.sun.security"/>
+    <allow pkg="org.glassfish.jersey"/>
+    <allow pkg="javax.ws.rs"/>
+    <allow pkg="org.glassfish.hk2"/>
+    <allow pkg="io.swagger"/>
+    <allow pkg="javax.validation"/>
+    <allow pkg="javax.crypto"/>
+    <allow pkg="com.fasterxml.jackson"/>
+    <allow pkg="mesosphere"/>
+    <allow pkg="feign"/>
+    <allow pkg="io.grpc" />
+    <allow pkg="info.ganglia" />
+    <allow pkg="org.gaul.s3proxy" />
+    <allow pkg="org.jclouds" />
+    <allow pkg="org.glassfish.grizzly" />
+    <allow pkg="com.spotify" />
+    <allow pkg="io.jsonwebtoken" />
+
+</import-control>
