+++ checkstyle/suppressions.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<!DOCTYPE suppressions PUBLIC
+        "-//Puppy Crawl//DTD Suppressions 1.1//EN"
+        "http://www.puppycrawl.com/dtds/suppressions_1_1.dtd">
+<suppressions>
+    <suppress checks="JavadocStyle" files=".+Tests\.java" />
+    <suppress checks="JavadocMethod" files=".+Tests\.java" />
+    <suppress checks="JavadocStyle" files=".+Test\.java" />
+    <suppress checks="JavadocMethod" files=".+Test\.java" />
+    <suppress files="[\\/]generated[\\/]" checks="[a-zA-Z0-9]*"/>
+    <suppress files="\.pem" checks="[a-zA-Z0-9]*"/>
+    <suppress files="passwd" checks="[a-zA-Z0-9]*"/>
+</suppressions>
+
