+++ gradle/checkstyle.gradle
@@ -1,0 +1,25 @@
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
+plugins.withId('checkstyle') {
+    checkstyle {
+        toolVersion = '8.2'
+
+        configFile = new File(rootDir, "checkstyle/checkstyle.xml")
+        configProperties = [importControlFile: "$rootDir/checkstyle/import-control.xml"]
+    }
+
+    plugins.withId('java') {
+        checkstyle {
+            // Exclude generated code from Checkstyle checks
+            checkstyleMain.exclude '**/rest/generated/**'
+        }
+    }
+}
