+++ gradle/eclipse.gradle
@@ -1,0 +1,20 @@
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
+def gradleProject = project
+plugins.withId('eclipse') {
+    if (gradleProject != gradleProject.rootProject) {
+        eclipse {
+            project {
+                name = "${gradleProject.rootProject.name}${gradleProject.path.replace(':', '-')}"
+            }
+        }
+    }
+}
