+++ gradle/idea.gradle
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
+plugins.withId('idea') {
+    idea.project?.ipr {
+        withXml { provider ->
+            // Add support for annotation processing which is needed by the lombok plugin
+            provider.node.component
+                    .find { it.@name == 'CompilerConfiguration' }
+                    .annotationProcessing.replaceNode {
+                        annotationProcessing {
+                            profile(name: 'Default', default: true, enabled: true)
+                        }
+                    }
+        }
+    }
+}
