+++ gradle/java.gradle
@@ -1,0 +1,85 @@
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
+plugins.withId('java') {
+    apply plugin: 'checkstyle'
+    apply plugin: 'findbugs'
+    apply plugin: 'jacoco'
+    apply plugin: 'signing'
+    apply plugin: 'maven'
+
+    compileJava {
+        options.compilerArgs.addAll([
+                "-Xlint:deprecation",
+                "-Xlint:divzero",
+                "-Xlint:empty",
+                "-Xlint:fallthrough",
+                "-Xlint:finally",
+                "-Xlint:overrides",
+                "-Xlint:path",
+                "-Werror"
+        ])
+    }
+
+    archivesBaseName = "pravega" + project.path.replace(':', '-')
+
+    task sourceJar(type: Jar) {
+        classifier = 'sources'
+        from sourceSets.main.java
+    }
+    artifacts { archives sourceJar }
+
+    task generateJavadoc(type: Javadoc) {
+        source = sourceSets.main.allJava
+        failOnError = false
+    }
+    task javadocJar(type: Jar) {
+        classifier = 'javadoc'
+        from generateJavadoc
+    }
+    artifacts { archives javadocJar }
+
+    task testJar(type: Jar) {
+        classifier = 'tests'
+        from sourceSets.test.output
+    }
+    artifacts { testRuntime testJar }
+
+    if (project.hasProperty("doSigning")) {
+        signing {
+            sign configurations.archives
+        }
+        // Default the secretKeyRingFile to the current user's home
+        if (!project.property("signing.secretKeyRingFile")) {
+            def secretKeyPath = project.file("${System.getProperty("user.home")}/.gnupg/secring.gpg").absolutePath
+            project.setProperty("signing.secretKeyRingFile", secretKeyPath)
+        }
+    }
+
+    tasks.withType(Test) {
+        systemProperties 'logback.configurationFile' : new File(buildDir,'resources/test/logback.xml').absolutePath
+        testLogging.showStandardStreams = false
+        testLogging.exceptionFormat = "FULL"
+        testLogging.showCauses = true
+        testLogging.showExceptions = true
+        testLogging.showStackTraces = true
+        testLogging.events = ["PASSED", "FAILED"]
+        maxParallelForks = System.properties['maxParallelForks'] ? System.properties['maxParallelForks'].toInteger() : 1
+        minHeapSize = "128m"
+        maxHeapSize = "512m"
+    }
+
+    dependencies {
+        // The production code uses the SLF4J logging API at compile time
+        compile group: 'org.slf4j', name:'slf4j-api', version: "1.7.25"
+        testCompile group: 'junit', name:'junit', version: "4.12"
+        testCompile "org.mockito:mockito-core:2.10.0"
+    }
+}
