+++ src/test/java/io/pravega/leaderelection/Client.java
@@ -6,8 +6,8 @@
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
+ *
  */
-
 package io.pravega.leaderelection;
 import lombok.extern.slf4j.Slf4j;
 import java.net.URI;
@@ -18,27 +18,20 @@ class Client extends Thread implements LeaderElection.LeaderElectionCallback {
     private static final String DEFAULT_SCOPE = "election";
     private static final String DEFAULT_CONFIG_NAME = "leaderElection";
     private static final String DEFAULT_CONTROLLER_URI = "tcp://127.0.0.1:9090";
-    private URI controllerURI;
-    private LeaderElection le;
-    private volatile boolean flag;
-
+    private final LeaderElection le;
 
     public Client(String hostName) {
-        controllerURI = URI.create(DEFAULT_CONTROLLER_URI);
+        URI controllerURI = URI.create(DEFAULT_CONTROLLER_URI);
         le = new LeaderElection(DEFAULT_SCOPE, DEFAULT_CONFIG_NAME, controllerURI,hostName, this);
-        flag = true;
     }
 
     @Override
     public void run() {
         le.start(500);
-        while(flag);
-        le.stop();
-        le.close();
     }
 
     public void stopRunning() {
-        flag = false;
+        le.stop();
     }
 
     public String get() {
