+++ src/test/java/io/pravega/leaderelection/LeaderElectionTest.java
@@ -6,39 +6,40 @@
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
+ *
  */
 package io.pravega.leaderelection;
-import junit.framework.TestCase;
+import org.junit.Assert;
 import org.junit.Test;
 import java.util.List;
 
 
-public class LeaderElectionTest extends TestCase {
+public class LeaderElectionTest {
     @Test
     public void testInitialElection3() throws InterruptedException {
         int host = 3;
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
+        Assert.assertNotNull(newLeader);
         cfg.stop();
     }
 
     @Test
     public void testMemberCrash3() throws InterruptedException {
         int host = 3;
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String leader = cfg.checkOneLeader();
-        assertNotNull(leader);
+        Assert.assertNotNull(leader);
 
         cfg.remain(leader);
         String leader2 = cfg.checkOneLeader();
-        assertNotNull(leader2);
-        assertEquals(leader2, leader);
+        Assert.assertNotNull(leader2);
+        Assert.assertEquals(leader2, leader);
 
         cfg.restartExcept(leader2);
         String leader3 = cfg.checkOneLeader();
-        assertNotNull(leader3);
-        assertEquals(leader3, leader2);
+        Assert.assertNotNull(leader3);
+        Assert.assertEquals(leader3, leader2);
 
         cfg.stop();
     }
@@ -46,18 +47,18 @@ public class LeaderElectionTest extends TestCase {
     @Test
     public void testReElection3() throws InterruptedException {
         int host = 2;
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String leader = cfg.checkOneLeader();
-        assertNotNull(leader);
+        Assert.assertNotNull(leader);
 
         cfg.disconnect(leader);
         String leader2 = cfg.checkOneLeader();
-        assertNotNull(leader2);
+        Assert.assertNotNull(leader2);
 
         cfg.connect(leader);
         String leader3 = cfg.checkOneLeader();
-        assertNotNull(leader3);
-        assertEquals(leader3, leader2);
+        Assert.assertNotNull(leader3);
+        Assert.assertEquals(leader3, leader2);
 
         cfg.stop();
     }
@@ -65,9 +66,9 @@ public class LeaderElectionTest extends TestCase {
     @Test
     public void testLeaderCorrectness3() throws InterruptedException {
         int host = 3;
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String leader = cfg.checkOneLeader();
-        assertNotNull(leader);
+        Assert.assertNotNull(leader);
 
         List<String> hosts = cfg.getAllConnectHost();
         hosts.remove(leader);
@@ -82,30 +83,30 @@ public class LeaderElectionTest extends TestCase {
 
         cfg.disconnect(leader);
         String newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
-        assertEquals(newLeader, h1);
+        Assert.assertNotNull(newLeader);
+        Assert.assertEquals(newLeader, h1);
         cfg.stop();
     }
 
     @Test
     public void testAllCrash3() throws InterruptedException {
         int host = 3;
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String leader = cfg.checkOneLeader();
-        assertNotNull(leader);
+        Assert.assertNotNull(leader);
 
         cfg.disconnect(leader);
         String newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
+        Assert.assertNotNull(newLeader);
 
         cfg.disconnect(newLeader);
         newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
+        Assert.assertNotNull(newLeader);
 
 
         cfg.disconnect(newLeader);
         newLeader = cfg.checkOneLeader();
-        assertNull(newLeader);
+        Assert.assertNull(newLeader);
 
         cfg.stop();
     }
@@ -114,9 +115,9 @@ public class LeaderElectionTest extends TestCase {
     public void testReElection5() throws InterruptedException {
         int host = 5;
 
-        TestConfig cfg = TestConfig.make_config(host);
+        TestGroup cfg = TestGroup.make_group(host);
         String leader = cfg.checkOneLeader();
-        assertNotNull(leader);
+        Assert.assertNotNull(leader);
 
         cfg.disconnect(leader);
         List<String> hosts = cfg.getAllConnectHost();
@@ -129,8 +130,8 @@ public class LeaderElectionTest extends TestCase {
         cfg.connect(hosts.get(2));
         cfg.connect(hosts.get(0));
         String newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
-        assertEquals(newLeader, hosts.get(3));
+        Assert.assertNotNull(newLeader);
+        Assert.assertEquals(newLeader, hosts.get(3));
 
         cfg.disconnect(newLeader);
         cfg.disconnect(hosts.get(1));
@@ -139,8 +140,8 @@ public class LeaderElectionTest extends TestCase {
         cfg.connect(hosts.get(1));
         cfg.connect(hosts.get(2));
         newLeader = cfg.checkOneLeader();
-        assertNotNull(newLeader);
-        assertEquals(newLeader, hosts.get(0));
+        Assert.assertNotNull(newLeader);
+        Assert.assertEquals(newLeader, hosts.get(0));
 
         cfg.stop();
     }
